2.4.1: greatsword's knights cannot spawn new knights

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Statue.java
@@ -140,8 +140,10 @@ public class Statue extends Mob {
 	
 	@Override
 	public void die( Object cause ) {
-		weapon.identify();
-		Dungeon.level.drop( weapon, pos ).sprite.drop();
+		if (weapon != null) {
+            weapon.identify();
+            Dungeon.level.drop(weapon, pos).sprite.drop();
+        }
 		super.die( cause );
 	}
